Rename the container to PixelClusterCountsInEvent
